requirements.txt
face_recognition==1.3.0
gradio==3.32.0
numpy==1.24.2
onnx==1.14.0
onnxruntime==1.14.1
Pillow==9.5.0
